Ensured results screen had a white background in both light and dark mode
This is a good reminder to ensure your layouts look the same/appropriate in light and dark mode. In this case, I had not defined the background colour of the results screen. This meant Android Studio used its default which was white for light mode and black for dark mode. This was a problem because the text on that screen was black.