Removed questionType parameter from QuestionTemp
Not necessary because when you create a quiz, you will specify the type of question you want. It is your job to remember what that type was. Also, it is bad design for every question to store this type, instead a class should contain the quiz and the question type, but that is unecessary.