Made correct option index selection more random
No two consecutive questions (in one quiz) have the answer in the same option index. For example, if the answer was in option index 2 for question 1, then question 2 could only contain the correct answer in option indices 0, 1, 3. This helps the player so they SHOULD NOT know this, however the increase in randomness is good. I also wrote a test.