Added names to the Infographics
After a long an painful 2 hours, will finally was able to figure out how to get all the TextViews into an array, it wasn't easy and stack overflow was little help intially, requiring the exploration of the deepest darkest hyperlinks google could provide. Additionally there is now an extra space between brown and kiwi, because im lazy and cant be bothered changing the size of everything to force the word kiwi to appear on the next line.